requirements.txt
pytorch-lightning==1.5.10
pyloudnorm==0.1.0
deep-phonemizer==0.0.17
https://github.com/pyannote/pyannote-audio/archive/develop.zip
soundfile
tqdm==4.62.3
